+++ src/001_config.js
@@ -1,1 +1,2 @@
 local.workerBootstrapUrl = 'worker.min.js';
+local.logAllExceptions = false;