+++ src/web/010_dispatch.js
@@ -49,10 +49,12 @@ local.web.dispatch = function dispatch(request) {
 		request.url = request.urld.protocol+'://'+request.urld.authority+request.urld.relative;
 	}
 
-	// Generate response
+	// Setup response object
+	var requestStartTime;
 	var response = new local.web.Response();
 	var response_ = local.promise();
 	response.on('headers', function() { processResponseHeaders(request, response); });
+	response.on('close', function() { response.latency = Date.now() - requestStartTime; });
 	if (request.stream) {
 		// streaming, fulfill on 'headers'
 		response.on('headers', function(response) {
@@ -75,6 +77,7 @@ local.web.dispatch = function dispatch(request) {
 	var args = Array.prototype.slice.call(arguments, 1);
 	args.unshift(function(request, response, schemeHandler) {
 		// execute by scheme
+		requestStartTime = Date.now();
 		schemeHandler = schemeHandler || local.web.schemes.get(scheme);
 		if (!schemeHandler) {
 			response.writeHead(0, 'unsupported scheme "'+scheme+'"');