+++ test/rtcpeer/tests.js
@@ -115,6 +115,7 @@ local.promise.bundle(responses_)
 	.always(function(responses) {
 		responses.forEach(function(res) {
 			print(res.body);
+			console.log(res.latency+' ms');
 		});
 		finishTest();
 	});
@@ -156,6 +157,7 @@ local.promise.bundle(responses_)
 	.always(function(responses) {
 		responses.forEach(function(res) {
 			print(res.body);
+			console.log(res.latency+' ms');
 		});
 		finishTest();
 	});